added OpenAPI support; added JWT verification on each critical endpoi…

+++ api/src/main/java/com/openclassrooms/dataShare_api/controller/FileController.java
@@ -10,6 +10,8 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -22,6 +24,9 @@ import org.springframework.web.multipart.MultipartFile;
 import com.openclassrooms.dataShare_api.dto.DSFileDTO;
 import com.openclassrooms.dataShare_api.service.FileService;
 
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import lombok.extern.slf4j.Slf4j;
 
 /**
@@ -41,10 +46,21 @@ public class FileController {
      * @param expirationDays how many days til it is expired (ie: 1 day = tomorrow)
      * @return [the stored file's unique identifier, 200 OK]; 400 BAD_REQUEST if an error occurred during the file's upload into the system; 500 INTERNAL_SERVER_ERROR otherwise
      */
+    @Operation(
+        summary = "Uploads a File"
+    )
+    @ApiResponses(value = {
+        @ApiResponse(responseCode = "200", description = "File uploaded"),
+        @ApiResponse(responseCode = "400", description = "File incorrectly sent"),
+        @ApiResponse(responseCode = "401", description = "Unauthorized"),
+        @ApiResponse(responseCode = "500", description = "Incorrect JWT provided")
+    })
     @PostMapping("/upload")
-    public ResponseEntity<?> upload(@RequestParam("file") MultipartFile file, @RequestParam("userId") String userId, @RequestParam("expirationDays") Long expirationDays) {
+    public ResponseEntity<?> upload(@RequestParam("file") MultipartFile file, @RequestParam("expirationDays") Long expirationDays, @AuthenticationPrincipal Jwt jwt) {
+        Long userId = jwt.getClaim("userId");
+        
         try {
-            String storedFilename = fileService.store(file, userId, expirationDays);
+            String storedFilename = fileService.store(file, String.valueOf(userId), expirationDays);
             log.info("[POST] /api/file/upload" + " [" + HttpStatus.OK +"]");
             return ResponseEntity.ok(Map.of("filename", storedFilename));
         } catch (RuntimeException e) {
@@ -62,8 +78,19 @@ public class FileController {
      * @param filename the file's unique identifier
      * @return the fetched File as a downloadable Resource
      */
-    @GetMapping("/download/{userId}/{filename}")
-    public ResponseEntity<Resource> download(@PathVariable Long userId, @PathVariable String filename) {
+    @Operation(
+        summary = "Downloads a File"
+    )
+    @ApiResponses(value = {
+        @ApiResponse(responseCode = "200", description = "File download should start"),
+        @ApiResponse(responseCode = "401", description = "Unauthorized"),
+        @ApiResponse(responseCode = "404", description = "File not found (or found but doesn't belong to User)"),
+        @ApiResponse(responseCode = "500", description = "Incorrect JWT provided")
+    })
+    @GetMapping("/download/{filename}")
+    public ResponseEntity<Resource> download(@PathVariable String filename, @AuthenticationPrincipal Jwt jwt) {
+        Long userId = jwt.getClaim("userId");
+
         try {
             Resource file = fileService.loadAsResource(userId, filename);
             log.info("[GET] /api/file/download/" + userId + "/" + filename + " [" + HttpStatus.OK +"]");
@@ -84,10 +111,21 @@ public class FileController {
      * @param fileId
      * @return the file's metadata as a DTO
      */
+    @Operation(
+        summary = "Fetches a File's metadata"
+    )
+    @ApiResponses(value = {
+        @ApiResponse(responseCode = "200", description = "File metadata found"),
+        @ApiResponse(responseCode = "401", description = "Unauthorized"),
+        @ApiResponse(responseCode = "404", description = "File not found (or found but doesn't belong to User)"),
+        @ApiResponse(responseCode = "500", description = "Incorrect JWT provided")
+    })
     @GetMapping("/{fileId}")
-    public ResponseEntity<DSFileDTO> getFile(@PathVariable Long fileId) {
+    public ResponseEntity<DSFileDTO> getFile(@PathVariable Long fileId, @AuthenticationPrincipal Jwt jwt) {
+        Long ownerId = jwt.getClaim("userId");
+
         try {
-            DSFileDTO fileData = fileService.getFileDTO(fileId);
+            DSFileDTO fileData = fileService.getFileDTO(fileId, ownerId);
             log.info("[GET] /api/file/" + fileId + " [" + HttpStatus.OK +"]");
             return ResponseEntity.ok(fileData);
         } catch (Exception e) {
@@ -101,9 +139,22 @@ public class FileController {
      * @param userId
      * @return the metadata of every file belonging to a given user as DTOs
      */
+    @Operation(
+        summary = "Fetches the metadata of every File belonging to a given User"
+    )
+    @ApiResponses(value = {
+        @ApiResponse(responseCode = "200", description = "Files' metadata found"),
+        @ApiResponse(responseCode = "401", description = "Unauthorized"),
+        @ApiResponse(responseCode = "500", description = "Incorrect JWT provided")
+    })
     @GetMapping("/list/{userId}")
-    public ResponseEntity<List<DSFileDTO>> listFiles(@PathVariable Long userId) {
+    public ResponseEntity<List<DSFileDTO>> listFiles(@PathVariable Long userId, @AuthenticationPrincipal Jwt jwt) {
+        Long ownerId = jwt.getClaim("userId");
+
         try {
+            if (!userId.equals(ownerId))
+                throw new Exception("Incorrect user token provided.");
+
             List<DSFileDTO> filesList = fileService.getFilesDTO(userId);
             log.info("[GET] /api/file/list/" + userId + " [" + HttpStatus.OK +"]");
             return ResponseEntity.ok(filesList);
@@ -118,9 +169,23 @@ public class FileController {
      * @param fileId
      * @return 200 OK if successful; 404 NOT_FOUND if the file is not found; 500 INTERNAL_SERVER_ERROR if an error occurred during the physical file's deletion
      */
+    @Operation(
+        summary = "Deletes a File"
+    )
+    @ApiResponses(value = {
+        @ApiResponse(responseCode = "200", description = "File deleted"),
+        @ApiResponse(responseCode = "401", description = "Unauthorized"),
+        @ApiResponse(responseCode = "404", description = "File not found (or found but doesn't belong to User)"),
+        @ApiResponse(responseCode = "500", description = "Incorrect JWT provided")
+    })
     @DeleteMapping("/{fileId}")
-    public ResponseEntity<HttpStatusCode> deleteFile(@PathVariable Long fileId) {
+    public ResponseEntity<HttpStatusCode> deleteFile(@PathVariable Long fileId, @AuthenticationPrincipal Jwt jwt) {
+        Long ownerId = jwt.getClaim("userId");
+
         try {
+            if (!ownerId.equals(fileService.getFileDTO(fileId, ownerId).getOwnerId()))
+                throw new Exception("You are not allowed to delete this file.");
+
             fileService.deleteFile(fileId);
             log.info("[DELETE] /api/file/" + fileId + " (" + HttpStatus.OK + ")");
             return new ResponseEntity<>(HttpStatus.OK);
